Add "faviconPath" parameter to site.json
If a website wants to use a favicon, it must specify the path of such an
icon inside the HTML's header (using <link>)[1].

Let's add a new setting "faviconPath" to the 'site.json' configuration
file. If the user specfies a path to an image, additional HTML code
will be inserted to use that image as the website's favicon (see
page.ejs).

[1] https://stackoverflow.com/a/7116701