favicon: fix missing starting slash if baseUrl is empty
We use a root-relative url path for favicon. The url path is prepended
with baseUrl.

If baseUrl is empty, then no starting slash would be prepended to the
url path, and the favicon will no longer be root-relative.

Let's fix the erroneous url path by always adding a starting slash, so
that even if baseUrl is empty, the url path will be prepended with '/'.