Page: Fix title prefix repeated on live reload
MarkBind modifies the title attribute of the page every time the page is
being loaded in order to prepend the title prefix. This results in the
modification being permanent.

When the page reloads, the attribute gets modified again and the prefix
gets prepended again. This results in multiple title prefix showing up
on the final re-rendered page.

Let's modify the rendering data's title (which is temporary, unlike the
attribute which is permanent) instead. The renderer will still render
the title with prefix correctly, while not making the title modification
permanent.