Fix broken baseUrl references for inner websites
The current working file path (ATTRIB_CWF) specifies the document that
is including another file (this included file's path is specified with
ATTRIB_INCLUDE_PATH).

When resolving references inside the included file, it should be
resolved with regards to the included file's directory, NOT the working
file's directory.

For example, if the current document is 'abc/index.md', and it includes
another document 'book/some-file.md', then resolving references in
'some-file.md' should be done with the 'book/' directory. not the 'abc/'
directory.

Let's use ATTRIB_INCLUDE_PATH instead of ATTRIB_CWF so that the baseUrl
is not resolved wrongly.