Site: Initialize baseUrl before user variables
The baseUrl variable is only initialized after all user variables have
been processed.

If the content of the user variables refer to baseUrl, baseUrl will be
evaluated as empty because it is not initialized yet.

Let's rearrange the order of variable initialization such that baseUrl
is initialized first. That way, user variables that reference it will
not have an empty value since it is initialized by the time the user
variables are processed.