Parser: Fix hostBaseUrl being replaced
Parser#_rebaseReference() may be called multiple times. The file content
may already have {{hostBaseUrl}} in it due to previous calls to
Parser#_rebaseReference().

The variable {{hostBaseUrl}} is still not defined at this stage.
Undefined variables are replaced by nunjucks with nothing,
resulting in wrong paths.

Let's prevent nunjucks from replacing {{hostBaseUrl}}, by declaring the
variable to be itself, so that the value can be injected later when it
actually exist.