Upgrade Bootstrap to v4.1.1
Let's update bootstrap.min.css to v4.1.1.

In order to also use "bootstrap.min.js", the new version of Bootstrap
also requires Popper.js and jQuery for certain components to work (Such
as for the displaying of Modals). See [1], under the Js section, "Show
components requiring JavaScript"

However, we are going to use bootstrap-vue in the next commit, which
discourages us from importing "bootstrap.min.js" if we are migrating a
project already using Bootstrap [2]. Furthermore, vue-strap already has
its own version of jQuery. Therefore, there is no need to import
"bootstrap.min.js" or its dependencies if we are going to remove it in
the next commit anyway.

[1] - https://getbootstrap.com/docs/4.1/getting-started/introduction/
[2] - https://bootstrap-vue.js.org/docs