Add temporary styles to prevent FOUC (#664)
The layout of the pages generated by MarkBind will appear broken at
first when the pages are first loaded in the browser. This is because
vue-strap has not been loaded yet, so the proper styles are not being
applied yet.

This causes an ugly page to be displayed for a few seconds before
vue-strap finishes loading. This is known as flash of unstyled content
(FOUC).

Let's try and fix FOUC by adding temporary styles, so that the layout
looks less broken before all of the loading is done.

We also removed the loading overlay in the commit, as it was the 
original solution for FOUC, but it results in the MarkBind generated
website perceived as slow.