Remove website-content wrapper from pages
All docs pages have a span named website-content wrapped around the main
body of content.

Nesting all content within a tag is an unnecessary complication.
The wrapper does not seem to add a significant value.

Let's remove the offending wrapper (and the corresponding css classes)
from everywhere in the site.