Add previous/next links to the user guide
Let's add previous/next links to the bottom of the main pages of the
user guide so that readers can read the relevant pages in the correct
sequence if they want to.

These links are not added to reference pages as they are not meant
to be read in a linear sequence.

A nunjucks macro is used to minimize duplicating boilerplate code.
Using a boilerplate include is not feasible because the reused code
has conditional contents. For example, the 'next' link should not
appear in the last page of the sequence.

`<variable name='title' id='title'>` syntax is used to make the
title of a page accessible from other pages. The duplicate id attribute
can be removed after MarkBind has added support for accessing page
variables from outside of the page.