Documentation: Remove unnecessary {{baseUrl}}
The documentation uses {{baseUrl}} to create absolute references to
many resources, including CSS files, links and images. This was
necessary because we did not have features to support dynamic resource
references. To ensure that our sites render consistently regardless of
the page that the user is in, we manually referenced each resource
with respect to an absolute {{baseUrl}}.

This is no longer necessary, as we can now support relative paths. As
Markbind can now automatically convert relative resource references to
absolute ones, we can remove {{baseUrl}} to make it easier to maintain
the documentation in the future.

In this commit, all unnecessary references to {{baseUrl}} have been
removed, except for mbdf files within userGuide/syntax and
userGuide/plugins. These files are meant as the central reference
within the user guide, and removing {{baseUrl}} from these files might
make it confusing to maintain or add new documentation in the future.