Add new mocks for fs-extra-promise#copyAsync
fs-extra-promise#copyAsync creates folders along the destination
if they do not already exist. For example:

fse.copyAsync('file.txt', 'does/not/exist/dest.txt')

will succeed and create the folders "does", "not" and "exist". This
is different from the behaviour of fs.copyFileSync, which raises an
error as the destination folder does not exist. Thus, our test mocks
for copyAsync are incorrect as they rely on the mock for copyFileSync.

Let's create new mocks for copyAsync and copyDirAsync such that their
behaviour matches that of the actual copyAsync.