Standardize variable declaration (#743)
We use <span> to declare variables in include and global variables.

<span> variables are difficult to differentiate from actual <span>
tag uses. Furthermore, in places where this differentiation is
important (e.g. pages where <span>s are extremely common), we are
forced to use <variable> tags, rather than <span> tags. This results in
two different (and confusing) ways of declaring variables.

Let's standardize all variable declaration to use <variable> instead
of <span>.

The processing of <span> variables are currently left intact, but
they are deprecated and should be removed in a subsequent release.