Fix regex for matching icon syntax
The current regex for matching icon syntax has the `m` multi-line flag,
which causes unexpected results in Markdown with special characters.
For example, consider the following Markdown source:

MON 1200 - 1400<br>
:fas-thumbs-down:<br>

1. markdown-it parses this source as a inline token, with `MON 1200 `
being parsed as text as it does not contain special characters that may
be part of new syntax [1].

2. markdown-it-regexp parses the remaining source, which contains
a match due to the `m` flag. However, it only replaces the first 17
characters (the length of `:fas-thumbs-down:`), thus replacing
`- 1400<br>\n:fas-t` in the source, which is incorrect.

Let's remove the `m` flag from the regex used to match icon syntax, and
add unit tests to ensure that icons in multi-line Markdown will be
parsed correctly in the future.

Without `m`, markdown-it-regexp will not match tokens beyond line
boundaries, and should only match icon syntax at the start of the
markdown-it parser source position, after all tokens before the icon
font have been parsed.

[1]: https://github.com/markdown-it/markdown-it/blob/d533c27017dbb9663422019a08adcf15265cf9fb/lib/rules_inline/text.js#L7-L10