Remove site nav expand button
The current functionality of the site nav expand button
is to change the width of site nav to either hide or show
it on smaller screen widths.

Under the new layout, changing the width of any side
navigation menus will result in content shifting within
the main content page. This is not reader friendly so let's
remove the functionality.