Update heading selector to support always-index and no-index
When using the built-in markbind search, authors are able to set
the heading level to be indexed.

We want to authors to be able to specify that certain headings should
be indexed or not be indexed regardless of the heading index level.

To support this, let's allow authors to add the {.always-index} or
{.no-index} attributes to headings to indicate that the heading should either
alwyas be indexed or not be indexed.