Use Sets instead of Dictionaries for Page.includedFiles
We use includedFiles to check what pages we need to rebuild when a page is updated.
When a file is included, we do something similar to `this.includedFiles[dependency.to] = true;`.
includedFiles is essentially a dictionary containing all {file: true} values.

Let's use the ES6 Set data structure to store this information. This allows us to use
page.includedFiles.has(filePath) when checking for existence, which is semantically
clearer (it clearly returns a boolean) as compared to page.includedFiles[filePath].