Parser: Fix crash if no href given to anchor or link tag
MarkBind crashes if no href is given to an anchor or link tag.

Let's introduce an extra check for the href attribute. Only if the
href attribute is present, do we proceed to rewrite URL to their
absolute paths.
Let's also include some additional functional tests to protect against
regression for this bug.