Asset: Update spacer-top class
The workaround for the spacer-top class is not required
as navbars do not have position:fixed with the latest changes.

However, the contents of the menus are too close to the top of
the page after "sticking" to the top. Thus the padding has been
changed to 1rem to follow Bootstrap's navbar spacing conventions.