DevDocs: Migrate to use MarkBind
Our Dev Guide is hosted on a wiki page on a separate repository,
and is difficult for contributors to locate.

The Dev Guide is a good way for us to demonstrate the capabilities
of MarkBind too, so let's feature our Dev Guide together with the
User Guide.

The "Docs" section of the existing MarkBind homepage is now renamed
to User Guide.