Add algolia-no-index class for hidden content (#785)
As MarkBind sites are rendered on the client side, the initial
HTML includes content that will be hidden to the reader after Vue
manipulates the DOM.

The DocSearch crawler indexes all content in the initial HTML,
including content that is hidden to the reader. This might be
confusing for the reader as he might not be able to find the indexed
content immediately on the page.

Let's add the algolia-no-index class to content that is known to be
hidden by VueStrap components. Authors can then add algolia-no-index
to the selectors_exclude attribute in the site's DocSearch
configuration [1] to exclude these content from being indexed.

[1] - https://community.algolia.com/docsearch/config-file.html#selectors_exclude-optional