Fix site nav padding
The padding in the site nav causes the title to be indented further
right than the list of contents, and creates a gap between the
scrollbar and the right side of the site nav.

Let's move the padding to the inner div that wraps the content
so that the padding is not applied for the scrollbar, and applied
for all content within the site nav.

Let's also update the bootstrap spacing classes for the list
elements inside the site nav to ensure that it is consistent
with the padding of other content in the site nav.