Developer Guide: Move to separate site from main user guide
As per [1], our dev docs should not be in the same folder as the
user guide. This is because user guides are only updated when
MarkBind is published, however dev docs are potentially updated
every time someone makes a change.

Let's create a separate site in /devdocs and move our Developer
Guide there. This entails a separate site entirely complete with
its own MarkBind files and site.json.
In doing so, we also need to create a separate repository for
travis to deploy to. These settings are reflected in
devdocs/site.json.

[1]: https://github.com/MarkBind/markbind/pull/793#issuecomment-478414584