header.md: Use absolute path instead of ../
Header contains link references such as ../index.html. This is so
that the links are correct when referring both the main site and
from a nested directory such as userGuide/. However, this approach
is incorrect if we have more deeply nested directories.

Instead of relative references, let's use absolute references w.r.t.
the root directory instead.

In addition, let's update the User Guide to let authors know that
they should always use absolute references in files within the
_markbind folder.