Parser: Support context-aware includes
After our parser includes a file, its context is discarded.
This means that in subsequent passes of parser, we do not know where
the content has been included from, and included content is treated
uniformly as if they were originally part of the host file.

The path of the references in the included file may not resolve
correctly if it belongs to a different directory from the host file.
For example, if /a/b/c.md refers to a file in '../file.css', and
/index.md is the host file that includes /a/b/c.md, then /index.md
should resolve this path to '/a/b/file.css'. It cannot retain the
original path '../file.css', as that will lead to the wrong css file.

To support context-aware references, i.e. the included segments
retain awareness of their original include locations, let's add
a data-included-from meta whenever we include files. To maintain
the existing behaviour and to prevent leaking file structure data,
this meta tag is removed before we finish generating the final
website.