Build/Deploy: Support building/deploying from multiple json configs
MarkBind only supports building and deploying from site.json

As per [1], it is useful to keep our user and developer guides
separate, as they are updated with different release cadences.
While the user guide should only be updated on release, the developer
guide needs to be up-to-date with each commit merged into master.

Let's teach MarkBind to support building and deploying from multiple
json configs, so that we can deploy different versions to different
sites.

Let's also add a new deployment command to our travis config to
trigger a deployment to the developer guide upon any commit to
the master branch.

[1]: https://github.com/MarkBind/markbind/pull/793#issuecomment-478414584