Plugins: Support shorthand syntax processing (#495)
Certain MarkBind elements such as panels require a lot of boilerplate 
HTML to be written in the .md file, which is tedious. We would like to 
have syntax shorthands for such element to make it easier for the 
site author, so that, for example, instead of writing:

    <panel>
        <span slot="header" class="card-title">
            Heading
        </span>
    </panel>

Authors can write this instead to achieve the same effect:

    <panel>
        <span heading>
            Heading
        </span>
    </panel>

Let's create a plugin that can convert shorthand syntaxes into proper 
MarkBind syntax simply, which serves as a framework for more of these
to be added.