update.bat: Fix missing /s for test_convert_sites' rmdir
When running test_convert_sites, we want to remove the
\non_markbind_site\_markbind folder and its content after
the update is complete, as the folder is generated by the
test and not supposed to be there originally.

The command responsible for it does not have the /s
parameter added, so it fails if the folder is not empty,
and \_markbind is never empty when generated.

Let's fix this by adding the /s parameter.