Fix site-nav.css not being added when layout contains site nav (#841)
When passing data to the page template, the siteNav variable only checks
for the existence of the site nav specified in the frontmatter. The
site-nav.css file is thus only added when the site nav is specified in
the frontmatter.

When a page has a layout with a site nav, but no site nav is specified
in the frontmatter, the site-nav.css file is not added, which causes the
styling of the site nav to break.

Let's update the logic when forming template data to check for the
existence of a site nav in the template when determining the value of
the siteNav variable.