Header: Update to selectively display links based on site
Header displays both the User Guide and Developer Guide regardless
of the site that the user is on.

We do not want to confuse the user with the Developer Guide on the
User Guide site, nor add unnecessary references to the User Guide on
the Developer Guide site.

Let's add environment conditions to these sites to selectively
display links on the header.