maintainerGuide.md: Move `npm version` step before updating test files
When building a website, MarkBind puts the version number in a meta tag
inside the generated HTML files. Therefore, functional tests' expected
files need to be updated, whenever there is a change in the version
number, so that the tests will continue to pass.

We asked maintainers to update the test files before the version number
is actually incremented, which doesn't make sense because the generated
HTML files will not have the new version number yet.

Let's re-arrange the steps so that the updating is done after the
version number is incremented.