Asset: remove icon .csv files
The files were used as a template to build the iconsMap
for using variable syntax to insert icons.

Since #680, we have shifted to another syntax for inserting
icons. We made the decision to depreciate the old
variable syntax. Thus it renders the current icon .csv files
as obsolete.

Let's remove these files as well when depreciating the
variable syntax for icons.