Update userGuide modal example to show actual code
The user guide did not show the lorem ipsum text,
differing from the displayed output.

Let's update the code such that both the example code
and the example demo code are the same.

Some ids have to be changed, since there are two modals
that have the same id, causing the wrong modal to be displayed.