Add thumbnails feature
To use an image as a thumbnail, we have to crop and rescale the image
by wrapping it in a div and using CSS. This can be especially
cumbersome with multiple images with different dimensions.

Let's add a built in feature that supports convenient creation of
thumbnails using a <thumb> tag. It only accepts one dimension attribute
`size`, rather than a separate `width` and `height`, and crops the
supplied image into a square or optionally a circle with minimal
hassle from the author.

Let's also extend that feature to also allow the author to supply text,
icons or emojis in place of an image.