Add PlantUML feature
Markbind page authors can use diagrams in their sites by linking them
in a `<pic>` tag. This is tedious as it requires the author to generate
the diagrams using an external tool, and the binary files generated by
that tool (e.g. .pptx, .psd) are not version control friendly.

Let's support PlantUML, a tool that allows users to write diagrams in
a simple and intuitive language. Users can then write their diagrams
in a text editor, and Markbind will generate the diagrams using
PlantUML, eliminating the need of an external editor.