Insert navbar temp styles after layout header
Previously, a fix was implemented for reducing FOUC by applying
temporary styles for <navbar> and <dropdown>. A later commit broke this
fix by moving the header and footer files to a layout, which caused
the contents of those files to be missed by the temporary style fix.

This commit fixes this by inserting the temporary styles after
the header and footer files are inserted in the page, such that
the temporary styles are applied correctly to the relevant tags.