Add importing variables from other files (#891)
Let's introduce a new syntax to import variables from other files:

<import
  varname1 varname2 ...
  from='filename.md'
  as='namespace'
/>

If varname1, varname2... attributes are specified,
the respective variables from filename.md can be used as variables
in the current page normally: {{varname1}} {{varname2}}.

If attribute 'as' is specified, ALL page variables from filename.md
can be accessed through the namespace specified in the attribute:
{{namespace.varname3}}, {{namespace.varname4}}

In extractPageVariables, we store all non-imported page variables into
a global lookup table. Also we detect what variables are imported ones,
and let nunjucks replace them with themselves first.
In the above example,
{{namespace.varname1}} would be replaced with {{namespace.varname1}}
in the first nunjucks call in includeFile after extractPageVariables.

After this, and before _preprocess, a new preprocessing step
is introduced that extracts the imported variables,
after which another nunjucks call is made so that the imported variables
can be rendered correctly.