Refactor page.ejs site template to be at src directory (#937)
Refactor page.ejs to be at src directory as discussed in #935. This is 
because template directory is hijacked to include markbind initialized 
templates.

In the near future, there seems to have no plans to have multiple base 
templates, as the design is very coupled with current base template in 
page.ejs. If there are plans to decouple, we should consider 
refactoring again.