Revert "Add language indicator to code blocks (#957)" (#964)
The language indicator may display the "incorrect" language, because we
sometimes use other language for syntax highlighting when the original
language isn't available (e.g. using "html" for MarkBind codes).

The language indicator also covers the first line of the source code if
it becomes too long.

Finally, this design may not be sustainable in the long run if we were
to add more features to code blocks (e.g. copy button).

This reverts commit c854937051f1222e4e68bf3659dc0b53196c4d8a.