Fix plugin live reloading for relative file paths
Relative file paths provided by plugins are not resolved against the
processed page or included file.

This results in live reloading failing for pages not in the root
directory.

Let's fix this, and also rewrite plugin source collection, allowing
plugins to simply define tags and corresponding source attributes
to watch.

This moves the logic for checking whether the source is part of an
included file to the main markbind repo, reducing potential code
duplication and simplifying the api.