Import markdown-it parser variant from /vue-strap
Markdown-it in /vue-strap uses different plugins for parsing
its markdown content inside component attributes.

In preparation for moving markdown-it parsing to /markbind, let's move
the markdown-it variant from /vue-strap over to maintain parity with the
past markdown rendering behaviour for attributes in vue components.
Let's also restructure the directory hierarchy for markdown-it in here,
to reduce code duplication between the two variants.