Implement per component markdown attribute parsing
Markdown passed in attributes are being rendered in the client.

This results in a dependency on the markdown parsing library
(markdown-it) on the client, bloating the resulting client bundle size.
It also slows webpage loading, especially on older devices.
Moreover, /markbind already parses some amount of markdown in attributes
for other purposes when building the site. This results in a less
cohesive codebase.

Let's move all markdown-it parsing from /vue-strap to /markbind,
passing the rendered html into vue slots.
Additionally, let's conveniently standardise the attribute naming for
these components at the same time.