Restructure generated site data for searchbar ui
Keywords tied to a heading are appended to the heading with a prepending
'|' character.
This is so that the '|' character would serve as a visual separator
between the heading and the keyword.

However, it is thus not possible to extract a keyword from its heading
deterministically, as a heading can also include the '|' character.

Let's move the keywords to its own field, allowing the searchbar to
extract a heading's keyword from it.