test/functional: Update expected files for added/removed blank lines
PR #990 removed (and unintentionally added) some blank lines. The
changes in the blank lines are not reflected in the tests' expected
files.

Our tests do not fail because it ignores whitespaces. Nevertheless, it
is still out of sync with the actual output, which means that anyone
running `npm run updatetest` or `npm run updatetestwin` will always see
these noisy changes.

Let's update the expected files to match the blank lines in the actual
output.