Fix lazy reload for urls without index
Lazy live reloading uses the file extension of the request url to
determine whether the request is for a html page, and not from a
dynamic include.
Hence, requests for index.html with urls that do not end with
index.html fail to be detected.

Let’s add extra processing to detect such requests.