Encapsulate anchor plugin logic and styles
The main markbind code has some amount of logic and styles relating to
the anchor plugin.

Let's encapsulate the anchor plugin's logic and styles within the
plugin itself, which should lead to better code readability.
This also allows the user to turn off the plugin without risk of side
effects.