Update variable documentation
The limitation of variables not being able to contain incomplete or
malformed xml has been patched.
Some areas of the variable documentation can also be formatted with
newly introduced features to be more visually appealing.

Let’s update the user guide to reflect the patch, and reformat some
areas of the variable documentation.