Fix site nav highlight and padding (#1242)
Needing to start content without padding in the site navigation menu is
a rare use case. Let's revert the padding removal from the site nav.

The site nav supports adding custom content in between navigation items
as a markdown list item, but erroneously applies the highlight and
click wrapper.
Let's not apply the wrapper when there is no anchor or nested list in
said list item to rectify this.