Update ignore configs for /vue-strap merge
Vue-strap enables coverage reporting for snapshot tests.

The linting and jest configurations need to be updated as well to
isolate linting and tests to their original repos as well.

Let's add the coverage folder to .gitignore to avoid large coverage
diffs. Let's also update the eslint, stylelint and jest configurations
to ignore the appropriate folders.

Some vue component no longer exist that are still in the .eslintrc file
as well.
Let's remove these.