Restructure logging utilities
@markbind/core's uses markbind-cli's logging utilities, which implies
it cannot function on its own.

The default logger instance as provided by winston and configured in
@markbind/core also includes the DailyRotateFile transport.
This causes unit tests for @markbind/core files that also mock the node
fs module to fail, since it is a file-based transport.

Let's restructure the logging slightly, providing @markbind/core with
it's own winston configuration without the DailyRotateFile transport.
Let's also increase the logging level of the console transport to
'debug', which was previously handled by the DailyRotateFile transport.
This ensures unit tests function properly, and allows @markbind/core to
function on its own.

Then, let's override this configuration in markbind-cli's
configuration, such that the old behaviour from a user standpoint is
preserved.