Fix puml plugin spawning zombie jvm processes (#1245)
Asynchronously executing the puml child process results in zombie jvm
processes as there are multiple callbacks attached to it and page
generation is heavily callback based.

Let's break page generation into chunks, which allows node to clean up
the zombie processes in a timely fashion.

Let's also change the generated image's path to resolve from the
location of it's definition (ie. the file where the <puml> tag is
used), to reduce the amount of jvms spawned through including the same
file multiple times.