Fix nunjucks path resolving
Nunjucks' features that use paths resolve from the specified template
directory by design.
In MarkBind's case, the template directories are the respective root
paths of the various sites and subsites of the project.

Hence, a separate nunjucks environment with the template directory
configured as such is needed for each site or subsite.

Let's introduce a wrapper abstraction for this, VariableRenderer, to
wrap over such a singular nunjucks environment.
The VariableRenderer instances will be managed by VariableProcessor
(previously VariablePreprocessor), which already has a framework set up
to render variables belonging to each site or subsite.

In addition, {% raw %} tags inside a layouts file do not work due to
two nunjucks passes being made.
Let's revert the removal of the helper functions to keep {% raw %} tags
in the first pass to fix this, while searching for a solution that
would allow a single pass for layouts.