Use tree-shaked BootstrapVue components
Only three components - modals, popovers and tooltips, rely on the
BootstrapVue library. The BootstrapVue dist is also rather large, at
358kb.

Let's tree shake the library and bundle it together with the
vue-components package as such.
This also makes for better cohesion of vue components.

In addition, the script tag referencing the polyfill dependency of
BootstrapVue and the components bundle was placed at the very bottom.
This increases FOUC, as the browser loads the bundles unnecessarily
after the entire page has loaded.

Let's move the script tags up to the <head> tag hence.