Add frontend live / hot reload workflow
The frontend workflow requires launching webpack in watch mode. This is
then followed by manually copying the built bundles to the output
directory of the site, or terminating and then relaunching the MarkBind
command.

With a frontend entry point setup in the core-web package, and the
MarkBind /vue-strap fork merged into the main monorepo, let's take
advantage of this and add a live / hot reload workflow.
This provides a near instant edit-feedback development cycle that makes
frontend development much quicker.

This is integrated into the markbind serve command using the --dev
option.