Force webpack build context to core-web directory
Webpack uses the current working directory as the build context from
which to resolve its loaders and modules,
and babel uses it as the conceptual project root directory.

When building frontend bundles, the working directory and thus build
context correctly resides at the core-web package.
When serving a site outside of the markbind/ project folder with the
--dev option however, it is not, causing the  core-web build process to
fail.

Let's explicitly set the build context for webpack and babel to the
core-web project folder hence.