Remove redundant Site unit tests
Several Site unit tests test assert the number of, and existence of
numerous assets and files.
These tests are rather unmaintainable, requiring frequent change
whenever any of these files are added, deleted, or moved.

Many hardcoded assets and files are also used in the mocked filesystem
for these tests, which are duplicates / mocks of the actual assets.
These may also need to be updated when the underlying assets change.

In addition, the cli functional tests, which are easily updated through
the npm scripts, are already able to detect the presence or absence of
such assets and files, the number of them, and any changes in the file
content.

Let's remove these unit tests as such.